# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Transport Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff members
      parameters:
        - name: hubId
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"

    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Staff created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{staffId}:
    get:
      operationId: getStaffMember
      tags: [staff]
      summary: Get staff member
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"

    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update staff member
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdate"
      responses:
        "200":
          description: Staff updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{staffId}/toggle-active:
    patch:
      operationId: toggleStaffActive
      tags: [staff]
      summary: Toggle staff active status
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"

  # ITEMS
...
  /reports/daily:
    get:
      operationId: getDailyReport
      tags: [reports]
      summary: Get daily report
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: hubId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
  /reports/hub-wise:
    get:
      operationId: getHubWiseReport
      tags: [reports]
      summary: Get hub-wise report
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Hub-wise report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HubReport"
  /reports/monthly:
    get:
      operationId: getMonthlyReport
      tags: [reports]
      summary: Get monthly report
      parameters:
        - name: months
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyReport"
  /reports/export:
    get:
      operationId: exportReport
      tags: [reports]
      summary: Export report as CSV
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: hubId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: CSV export URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"

  # SEARCH
...
components:
  schemas:
...
    StaffUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
    StaffInput:
      type: object
      required: [name, phone, email, password, role]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
    StaffMember:
      type: object
      required: [id, name, phone, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
...
    DailyReport:
      type: object
      required: [date, totalBooked, totalDelivered, totalPending, revenueToday]
      properties:
        date:
          type: string
        totalBooked:
          type: integer
        totalDelivered:
          type: integer
        totalPending:
          type: integer
        revenueToday:
          type: number
    HubReport:
      type: object
      required: [hubId, hubName, hubCode, bookings, deliveries, pending, revenue, complaints]
      properties:
        hubId:
          type: integer
        hubName:
          type: string
        hubCode:
          type: string
        bookings:
          type: integer
        deliveries:
          type: integer
        pending:
          type: integer
        revenue:
          type: number
        complaints:
          type: integer
    MonthlyReport:
      type: object
      required: [month, year, totalParcels, delivered, deliveryPercentage, complaintRate]
      properties:
        month:
          type: integer
        year:
          type: integer
        label:
          type: string
        totalParcels:
          type: integer
        delivered:
          type: integer
        deliveryPercentage:
          type: number
        complaintRate:
          type: number
        revenue:
          type: number
    ExportResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
...